+++ test/fixtures/namespace-members-project/src/index.ts
@@ -1,0 +1,4 @@
+import { Tools, FullyUsed } from "./lib";
+
+Tools.used();
+console.log(FullyUsed);